Use different nodes comparison method
