+++ inference-engine/tests/functional/plugin/CMakeLists.txt
@@ -20,7 +20,6 @@ if (ENABLE_MYRIAD)
     add_subdirectory(myriad)
 endif()
 
-if (ENABLE_CONFORMANCE_TESTS)
-    add_subdirectory(conformance)
-endif()
+
+add_subdirectory(conformance)
