+++ inference-engine/tests/functional/plugin/conformance/CMakeLists.txt
@@ -3,7 +3,5 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-if (ENABLE_CONFORMANCE_TESTS)
-    add_subdirectory(subgraphs_dumper)
-    add_subdirectory(subgraphs_dumper/tests)
-endif()
+add_subdirectory(subgraphs_dumper)
+add_subdirectory(subgraphs_dumper/tests)
