+++ inference-engine/tests/functional/plugin/conformance/subgraphs_dumper/src/main.cpp
@@ -15,47 +15,50 @@
 int main(int argc, char *argv[]) {
     uint8_t ret_code = 0;
 
-    try {
-        gflags::ParseCommandLineNonHelpFlags(&argc, &argv, true);
-        if (FLAGS_h) {
-            showUsage();
-            return 0;
-        }
-        auto ie = InferenceEngine::Core();
 
-        auto cache = SubgraphsDumper::OPCache::make_cache();
+    gflags::ParseCommandLineNonHelpFlags(&argc, &argv, true);
+    if (FLAGS_h) {
+        showUsage();
+        return 0;
+    }
+    auto ie = InferenceEngine::Core();
 
-        if (!CommonTestUtils::directoryExists(FLAGS_input_folders)) {
-            std::string msg = "Input directory (" + FLAGS_input_folders + ") is not exist!";
-            throw std::runtime_error(msg);
-        }
+    auto cache = SubgraphsDumper::OPCache::make_cache();
 
-        if (!CommonTestUtils::directoryExists(FLAGS_output_folders)) {
-            std::string msg = "Output directory (" + FLAGS_output_folders + ") is not exist!";
-            throw std::runtime_error(msg);
-        }
+    if (!CommonTestUtils::directoryExists(FLAGS_input_folders)) {
+        std::string msg = "Input directory (" + FLAGS_input_folders + ") is not exist!";
+        throw std::runtime_error(msg);
+    }
 
-        std::vector<std::string> input_folder_content;
-        CommonTestUtils::directoryFileListRecursive(FLAGS_input_folders, input_folder_content);
+    if (!CommonTestUtils::directoryExists(FLAGS_output_folders)) {
+        std::string msg = "Output directory (" + FLAGS_output_folders + ") is not exist!";
+        throw std::runtime_error(msg);
+    }
 
-        for (const auto &file : input_folder_content) {
+    std::vector<std::string> input_folder_content;
+    CommonTestUtils::directoryFileListRecursive(FLAGS_input_folders, input_folder_content);
+
+    for (const auto &file : input_folder_content) {
+        try {
             if (CommonTestUtils::fileExists(file) && std::regex_match(file, std::regex(R"(.*\.xml)"))) {
                 std::cout << "Processing model: " << file << std::endl;
                 std::string bin_file = CommonTestUtils::replaceExt(file, "bin");
                 if (!CommonTestUtils::fileExists(bin_file)) {
-                    std::cerr << "Corresponding .bin file for the model " << file <<  " doesn't exist" << std::endl;
+                    std::cerr << "Corresponding .bin file for the model " << file << " doesn't exist" << std::endl;
                     continue;
                 }
                 InferenceEngine::CNNNetwork net = ie.ReadNetwork(file);
                 auto function = net.getFunction();
                 cache->update_ops_cache(function, file);
             }
+        } catch (std::exception &e) {
+            std::cerr << "Model processing failed with exception:" << std::endl << e.what() << std::endl;
+            ret_code = 2;
+            continue;
         }
-
-        cache->serialize_cached_ops(FLAGS_output_folders);
-    } catch (std::exception &e) {
-        std::cerr << "Processing failed with exception:" << std::endl << e.what() << std::endl;
-        ret_code = 2;
     }
+
+    cache->serialize_cached_ops(FLAGS_output_folders);
+
     return ret_code;
 }
