+++ inference-engine/tests/functional/plugin/conformance/subgraphs_dumper/src/matchers/single_op.cpp
@@ -49,6 +49,9 @@ bool compare_constants_data(const std::shared_ptr<ngraph::op::Constant> &op,
             return compare_constants_data<int32_t>(op, ref);
         case ngraph::element::Type_t::i64:
             return compare_constants_data<int64_t>(op, ref);
+        // TODO cast_vector doesn't support u1 now
+//        case ngraph::element::Type_t::u1:
+//            return compare_constants_data<char>(op, ref);
         case ngraph::element::Type_t::u8:
             return compare_constants_data<uint8_t>(op, ref);
         case ngraph::element::Type_t::u16:
@@ -58,7 +61,8 @@ bool compare_constants_data(const std::shared_ptr<ngraph::op::Constant> &op,
         case ngraph::element::Type_t::u64:
             return compare_constants_data<uint64_t>(op, ref);
         default:
-            throw std::runtime_error("unsupported type");
+            std::cerr << "Can't compare constants" << op << " with " << ref << "\n" << "Unsupported data type";
+            return false;
     }
 }
 
@@ -71,7 +75,9 @@ bool SingleOpMatcher::match(const std::shared_ptr<ngraph::Node> &node, const std
         return false;
     }
     // Match inputs size
-    if (node->get_input_size() == ref->get_input_size()) {
+    if (node->get_input_size() != ref->get_input_size()) {
+        return false;
+    } else {
         // Match input ranks, element types and static shapes
         for (size_t i = 0; i < node->get_input_size(); ++i) {
             bool rankIsEqual = node->get_input_tensor(i).get_partial_shape().rank() ==
@@ -84,12 +90,12 @@ bool SingleOpMatcher::match(const std::shared_ptr<ngraph::Node> &node, const std
                 return false;
             }
         }
-    } else {
-        return false;
     }
 
     // Match outputs size
-    if (node->get_output_size() == ref->get_output_size()) {
+    if (node->get_output_size() != ref->get_output_size()) {
+        return false;
+    } else {
         // Match output element type
         for (size_t i = 0; i < node->get_output_size(); ++i) {
             if (node->get_output_tensor(i).get_element_type() !=
@@ -97,23 +103,15 @@ bool SingleOpMatcher::match(const std::shared_ptr<ngraph::Node> &node, const std
                 return false;
             }
         }
-    } else {
-        return false;
     }
     // Compare node attributes
-    CompareNodesAttributes attrs_comparator;
-    auto &node_visitor = attrs_comparator.get_cmp_reader();
-    auto &ref_visitor = attrs_comparator.get_ref_reder();
-    // TODO: Need to apply readers wise-versa, otherwise comparison fails, why?
-    node->visit_attributes(ref_visitor);
-    ref->visit_attributes(node_visitor);
-    // TODO: Implement equal() with attributes mask - which attributes to compare and which not
-    if (!attrs_comparator.equal()) {
+    auto equal = attributes::compare(node.get(), ref.get(), Comparator::CmpValues::ATTRIBUTES);
+    if (!equal.valid) {
         return false;
     }
     // Match ports values
     const auto &cfg = get_config(node);
-    std::vector<size_t> ignored_ports = cfg->ignored_ports;
+    const std::vector<size_t> &ignored_ports = cfg->ignored_ports;
 
     for (size_t port_id = 0; port_id < node->get_input_size(); ++port_id) {
         if (std::find(ignored_ports.begin(), ignored_ports.end(), port_id) != ignored_ports.end()) {
@@ -152,8 +150,6 @@ SingleOpMatcher::SingleOpMatcher() {
                     ngraph::op::v1::Add,
                     ngraph::op::v1::Multiply,
                     ngraph::op::v1::Subtract,
-                    ngraph::op::v1::Power,
-                    ngraph::op::v1::ReduceMax,
-                    ngraph::op::v1::ReduceMin>>(std::vector<std::string>{}, std::vector<size_t>{0, 1}),
+                    ngraph::op::v1::Power>>(std::vector<std::string>{}, std::vector<size_t>{0, 1}),
     };
 }
